==> 01_Set Up.ipynb <==
# SET UP

## CREAR EL ENTORNO PARA EL PROYECTO

Crear un nuevo entorno e instalar los paquetes básicos:

1. Ir al prompt o terminal
2. Copiar y poner la línea inferior:

original
conda create --name proyecto1 python numpy pandas matplotlib seaborn scikit-learn scipy sqlalchemy xgboost jupyter

mia
conda create --name proyecto1 -c conda-forge python numpy pandas matplotlib seaborn scikit-learn scipy sqlalchemy xgboost jupyter

3. Activar el nuevo entorno:

conda activate proyecto1

4. Instalar los paquetes que están en otros canales:

conda install -c conda-forge pyjanitor scikit-plot yellowbrick imbalanced-learn jupyter_contrib_nbextensions cloudpickle

conda install -c districtdatalabs yellowbrick

pip install category_encoders

pip install notebook==7.3.0
pip install jupyter-client==8.6.3

5. Crear el environment.yml

conda env export > proyecto1.yml



## IMPORTAR PAQUETES

import os
import numpy as np
import pandas as pd

#Automcompletar rápido
%config IPCompleter.greedy=True

## CREAR EL DIRECTORIO DEL PROYECTO

### Definir el directorio raiz donde vamos a crear el proyecto

raiz = 'C:/Users/alvar/Proyectos ML/'

### Definir el nombre del proyecto

nombre_dir = '03_RIESGOS'

### Crear el directorio y la estructura del proyecto

path = raiz + nombre_dir

path = raiz + nombre_dir

try:
    os.mkdir(path)
    os.mkdir(path + '/01_Documentos')
    os.mkdir(path + '/02_Datos')
    os.mkdir(path + '/02_Datos/01_Originales')
    os.mkdir(path + '/02_Datos/02_Validacion')
    os.mkdir(path + '/02_Datos/03_Trabajo')
    os.mkdir(path + '/02_Datos/04_Caches')
    os.mkdir(path + '/03_Notebooks')
    os.mkdir(path + '/03_Notebooks/01_Funciones')
    os.mkdir(path + '/03_Notebooks/02_Desarrollo')
    os.mkdir(path + '/03_Notebooks/03_Sistema')
    os.mkdir(path + '/04_Modelos')
    os.mkdir(path + '/05_Resultados')
    os.mkdir(path + '/09_Otros')
    
except OSError:
    print ("La creación del directorio %s ha fallado" % path)
else:
    print ("Se ha creado satisfactoriamente el directorio %s " % path)

### Situar el directorio de trabajo en el proyecto

os.chdir(path)

### Mover el environment.yml

Buscar manualmente el archivo proyecto1.yml, que posiblemente lo haya ubicado en C:\Users\tu_usuario y moverlo a la carpeta /01_Documentos'

## CREAR LOS DATASETS INICIALES

Poner los datos originales en la carpeta '/02_Datos/01_Originales'

### Importar los datos

Sustituir el nombre_fichero_datos.

nombre_fichero_datos = 'prestamos.csv'

Cargar los datos.

ruta_completa = path + '/02_Datos/01_Originales/' + nombre_fichero_datos

datos = pd.read_csv(ruta_completa,index_col=0)
datos

### Extraer y reservar el dataset de validación

val= datos.sample(frac = 0.3)

nombre_fichero_validacion = 'validacion_mio.csv'

ruta_completa = path + '/02_Datos/02_Validacion/' + nombre_fichero_validacion

val.to_csv(ruta_completa)

### Extraer y guardar el dataset de trabajo

trabajo = datos.loc[~datos.index.isin(val.index)]

nombre_fichero_trabajo = 'trabajo_mio.csv'

ruta_completa = path + '/02_Datos/03_Trabajo/' + nombre_fichero_trabajo

trabajo.to_csv(ruta_completa)

### Extraer y guardar una muestra (opcional)

Actualizar el tamaño de la muestra.

muestra= trabajo.sample(n = 20000)

nombre_fichero_muestra = 'muestra.csv'

ruta_completa = path + '/02_Datos/03_Trabajo/' + nombre_fichero_muestra

muestra.to_csv(ruta_completa)
