==> 04_Transformacion de datos.ipynb <==
# TRANSFORMACIÓN DE DATOS

## IMPORTAR PAQUETES

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder

from sklearn.preprocessing import KBinsDiscretizer
from sklearn.preprocessing import Binarizer

from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.preprocessing import MinMaxScaler

#Automcompletar rápido
%config IPCompleter.greedy=True

## IMPORTAR LOS DATOS

1.- Sustituir la ruta del proyecto.

ruta_proyecto = 'C:/Users/alvar/Proyectos ML/03_RIESGOS'

2.- Nombrar los ficheros de datos.

nombre_cat = 'cat_resultado_eda.pickle'
nombre_num = 'num_resultado_eda.pickle'

3.- Cargar los datos.

cat = pd.read_pickle(ruta_proyecto + '/02_Datos/03_Trabajo/' + nombre_cat)
num = pd.read_pickle(ruta_proyecto + '/02_Datos/03_Trabajo/' + nombre_num)

Recordamos lo que habíamos identificado en fases anteriores y lo organizamos por tipo de acción.

Modificar con pandas/numpy:

* Crear la target a partir de estado
* Discretizar num_meses_desde_ult_retraso
* En vivienda juntar NONE y OTHER
* En finalidad juntar renewable_energy, educational y vacation en 'otros'

Modificar con scikit learn:

* La variable descripcion podríamos trabajarla con un TF-IDF Vectorizer
* Convertir num_derogatorios a binaria

## CREACIÓN DE VARIABLES Y TRANSFORMACIONES CON PANDAS

### Target para el modelo EAD

num.info()

num['target_ead']=num.apply(lambda x: ((x['principal']-x['imp_amortizado'])/x['principal']) if ((x['principal']-x['imp_amortizado'])/x['principal'])>=0 else 0 ,axis=1)


num['target_ead'].describe().T


### Target para el modelo LGD

num.info()

num['target_lgd']=num.apply(lambda x: (1-(x['imp_recuperado']/(x['principal']-x['imp_amortizado'])))
                        if (1-(x['imp_recuperado']/(x['principal']-x['imp_amortizado'])))>=0 else 0,axis=1)

num['target_lgd'].isna().sum()

Comprobamos las targets creadas.

cat['target_pd'].mean()

num[['target_ead','target_lgd']].describe().T

Eliminamos las variables que crearon la target

num.drop(columns=['imp_recuperado','imp_amortizado'],inplace=True)

num.drop(columns='num_meses_desde_ult_retraso',inplace=True)

## TRANSFORMACIÓN DE CATEGÓRICAS

cat.columns.to_list()

### One Hot Encoding

#### Variables a aplicar OHE

var_ohe = [ 'ingresos_verificados', 'vivienda','finalidad','num_cuotas']

#### Instanciar

ohe = OneHotEncoder(sparse= False, handle_unknown='ignore')

#### Entrenar y aplicar

cat_ohe = ohe.fit_transform(cat[var_ohe])

#### Guardar como dataframe

cat_ohe = pd.DataFrame(cat_ohe, columns = ohe.get_feature_names_out())

### Ordinal Encoding

#### Variables a aplicar OE

var_oe = ['antiguedad_empleo','rating']

#### Orden de los valores de las variables

#Orden de la primera variable
orden_antiguedad_empleo = ['desconocido','< 1 year','1 year','2 years','3 years','4 years',
                           '5 years','6 years','7 years','8 years','9 years','10+ years']

#Orden de la segunda variable
orden_rating =['G','F','E','D','C','B','A']

#### Instanciar

oe = OrdinalEncoder(categories = [orden_antiguedad_empleo,orden_rating],
                    handle_unknown = 'use_encoded_value',
                    unknown_value = 12)

#### Entrenar y aplicar

cat_oe = oe.fit_transform(cat[var_oe])

#### Guardar como dataframe

#Añadir sufijos a los nombres
nombres_oe = [variable + '_oe' for variable in var_oe]

#Guardar como dataframe
cat_oe = pd.DataFrame(cat_oe, columns = nombres_oe)

cat_oe

### Fechas y textos

Cargamos stopwords y limpiamos

stop = pd.read_csv(ruta_proyecto + '/01_Documentos/stop_words_english.txt',names = ['termino'])

def quitar_tildes(palabra):
    #Definimos la versión con tildes y símbolos y la sin
    con = 'áéíóúüñÁÉÍÓÚÜÑ'
    sin = 'aeiouunAEIOUUN'
    #Creamos un traductor
    traductor = str.maketrans(con,sin)
    #Aplicamos el traductor y devolvemos la palabra limpia
    return(palabra.translate(traductor))

stop['limpias'] = stop.termino.transform(quitar_tildes)

Aplicamos tf-idf.

#Instanciamos
tfidf = TfidfVectorizer(strip_accents = 'unicode',
                     stop_words = stop.limpias.to_list(),
                     max_df = 0.7,
                     min_df = 50,
                     ngram_range = (1,3),
                     max_features = 50)

#Entrenamos y aplicamos
descripcion = tfidf.fit_transform(cat.descripcion)

tfidf.vocabulary_

No parece que estos términos puedan generar información adicional a lo que tenemos, así que no incluiremos esta variable en los modelos.

#Eliminamos de cat las variables que no vamos a usar
cat=cat.drop(columns=['empleo','descripcion'])

## TRANSFORMACIÓN DE NUMÉRICAS

num.describe().T

num.columns.to_list()

## REESCALAR VARIABLES

### Con Min-Max

num['num_derogatorios']


#df de las variables que no sufren transformacion

df_otros = pd.DataFrame({
    'num_cancelaciones_12meses': num['num_cancelaciones_12meses'],
    'num_derogatorios': num['num_derogatorios']
}, index=num.index)


Unificar los datasets a reescalar.

num_escalar = num[['ingresos',
                  'dti',
                  'num_lineas_credito',
                  'porc_uso_revolving',
                  'principal',
                  'tipo_interes',
                  'imp_cuota',
                  'num_hipotecas',
                  'porc_tarjetas_75p']].reset_index(drop=True)

df_res = pd.concat([cat_oe,num_escalar], axis=1)

#### Variables a reescalar con Min-Max

var_mms = df_res.columns

#### Instanciar

mms = MinMaxScaler()

#### Entrenar y aplicar

df_mms = mms.fit_transform(df_res[var_mms])

#### Guardar como dataframe

#Añadir sufijos a los nombres
nombres_mms = [variable + '_mms' for variable in var_mms]

#Guardar como dataframe
df_mms = pd.DataFrame(df_mms,columns = nombres_mms)

## UNIFICAR DATASETS

### Crear una lista con los dataframes y variables a incluir en el tablón analítico para el modelo de PD

incluir_pd = [cat_ohe, df_mms, df_otros.reset_index().drop(columns=['id_cliente']), cat.target_pd.reset_index()]

### Unir todos los dataframes en el tablón analítico para el modelo de PD

df_pd = pd.concat(incluir_pd, axis = 1)

df_pd.set_index('id_cliente',inplace=True)

### Crear una lista con los dataframes y variables a incluir en el tablón analítico para el modelo de EAD

incluir_ead = [cat_ohe, df_mms, df_otros.reset_index().drop(columns=['id_cliente']), num.target_ead.reset_index()]

### Unir todos los dataframes en el tablón analítico para el modelo de EAD

df_ead = pd.concat(incluir_ead, axis = 1)

df_ead.set_index('id_cliente',inplace=True)

### Crear una lista con los dataframes y variables a incluir en el tablón analítico para el modelo de LGD

incluir_lgd = [cat_ohe, df_mms, df_otros.reset_index().drop(columns=['id_cliente']), num.target_lgd.reset_index()]

### Unir todos los dataframes en el tablón analítico para el modelo de LGD

df_lgd = pd.concat(incluir_lgd, axis = 1)

df_lgd.set_index('id_cliente',inplace=True)

## GUARDAR DATASETS TRAS TRANSFORMACIÓN DE DATOS

En formato pickle para no perder las modificaciones de metadatos.

#Definir los nombres de los archivos
ruta_df_tablon_pd = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'df_tablon_pd.pickle'
ruta_df_tablon_ead = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'df_tablon_ead.pickle'
ruta_df_tablon_lgd = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'df_tablon_lgd.pickle'

# #Guardar los archivos
df_pd.to_pickle(ruta_df_tablon_pd)
df_ead.to_pickle(ruta_df_tablon_ead)
df_lgd.to_pickle(ruta_df_tablon_lgd)

df_lgd.shape
